i and j keys no longer scroll in results.html https://bugs.webkit.org/show_bug.cgi?id=92584
Reviewed by Tony Chang.

Scrolling broke with r122880 since the scrolling element is no
longer the body.

Also, fix TestNavigator to allow navigating to tests in the
unexpected passes table. That table doesn't have expand links,
so we change TestNavigator to navigate the tbody's instead, which
also happens to remove a bunch of unnecessary tree-walking.

* fast/harness/resources/results-test.js:
* fast/harness/results-expected.txt:
* fast/harness/results.html:


Identifier: 110662@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@124287 268f45cc-cd09-0410-ab3c-d52691b4dbfc