Regression(r124564): Wrong inlineChildrenBlock->hasLayer() computed in RenderBlock::removeChild. https://bugs.webkit.org/show_bug.cgi?id=90800
Reviewed by Eric Seidel.

r124564 reversed the sequence of setStyle and removeChildNode calls, but failed to cache the value
of inlineChildrenBlock->hasLayer(). So, it will be null when the layer is removed from parent in setStyle.
Fixed by the caching the bool value.

Covered by existing test fast/block/layer-not-removed-from-parent-crash.html.

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::removeChild):

Identifier: 110926@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@124580 268f45cc-cd09-0410-ab3c-d52691b4dbfc