If the Apple Java plug-in is blocked and no runtime is installed, don't load it https://bugs.webkit.org/show_bug.cgi?id=93289 <rdar://problem/11730092>
Reviewed by Dan Bernstein.

Source/WebKit2:

If the Apple Java plug-in is blocked, but there's no Java runtime installed (or the Java plug-in is disabled),
don't even include it in the plug-in info store.

* UIProcess/Plugins/PluginInfoStore.cpp:
(WebKit::PluginInfoStore::shouldBlockPlugin):
Since this is static now, it shouldn't be const.

* UIProcess/Plugins/PluginInfoStore.h:
(PluginInfoStore):
shouldBlockPlugin can be static.

* UIProcess/Plugins/mac/PluginInfoStoreMac.mm:
(WebKit::PluginInfoStore::shouldUsePlugin):
Return false if the Apple Java plug-in is blocked but Java isn't installed or enabled.

(WebKit::PluginInfoStore::shouldBlockPlugin):
Since this is static now, it shouldn't be const.

WebKitLibraries:

Add WKJLIsRuntimeAndWebComponentsInstalled().

* WebKitSystemInterface.h:
* libWebKitSystemInterfaceLion.a:
* libWebKitSystemInterfaceMountainLion.a:

Identifier: 111134@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@124799 268f45cc-cd09-0410-ab3c-d52691b4dbfc