[Chromium] Skia assert triggered by custom filter unmultiplied results https://bugs.webkit.org/show_bug.cgi?id=92758
Reviewed by Dean Jackson.

Source/WebCore:

The FECustomFilter (which renders the CSS Shaders) used premultiplied colors for both input and ouput.
The problem with that is that some shaders do not pre-multiply the values and the reality is that
we cannot trust the shader to do that. The easy way would be to use only unpremultiplied colors and do the
computation in C++ code. This will be changed when the mix blending is implemented to do that in the shader itself.

No new tests, just updated existing tests to take unpremultiplied colors instead.

* platform/graphics/filters/FECustomFilter.cpp:
(WebCore::FECustomFilter::platformApplySoftware):

LayoutTests:

Updated the shaders to output un-premultiplied colors. Also, updated the output results where needed.

* css3/filters/custom/custom-filter-shader-cache-expected.png:
* css3/filters/custom/effect-color-check-expected.png:
* css3/filters/custom/effect-custom-combined-missing-expected.png:
* css3/filters/custom/effect-custom-expected.png:
* css3/filters/custom/effect-custom-parameters-expected.png:
* css3/filters/custom/filter-repaint-custom-expected.png:
* css3/filters/custom/missing-custom-filter-shader-expected.png:
* css3/filters/resources/color-add.fs:
* css3/filters/resources/color-fill.fs:
* platform/chromium/TestExpectations: Disabled the following tests on Windows. They need a rebaseline.
* platform/chromium-linux/css3/filters/custom/effect-color-check-expected.png:
* platform/chromium-linux/css3/filters/custom/effect-custom-combined-missing-expected.png:
* platform/chromium-linux/css3/filters/custom/effect-custom-parameters-expected.png:
* platform/chromium-mac/css3/filters/custom/effect-color-check-expected.png:
* platform/chromium-mac/css3/filters/custom/effect-custom-combined-missing-expected.png:
* platform/chromium-mac/css3/filters/custom/effect-custom-parameters-expected.png:

Identifier: 111352@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125044 268f45cc-cd09-0410-ab3c-d52691b4dbfc