REGRESSION (r123171): <svg> element with intrinsic size and max-width gets sized incorrectly https://bugs.webkit.org/show_bug.cgi?id=93388
Reviewed by Beth Dakin.

Source/WebCore:

Test: svg/css/max-width-3.html

* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::computeAspectRatioInformationForRenderBox): Added code to set
m_intrinsicSize in the no-contentRenderer branch of this function so that the calls to
RenderBox::computeReplacedLogical{Height,Width} in the end use the right intrinsic size.
This is similar to what the contentRenderer branch of this function already does.

LayoutTests:

* svg/css/max-width-3-expected.html: Added.
* svg/css/max-width-3.html: Added.



Identifier: 111358@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125050 268f45cc-cd09-0410-ab3c-d52691b4dbfc