Avoid backing store on layers created for CoreAnimation plugins https://bugs.webkit.org/show_bug.cgi?id=93526 <rdar://problem/12052828>
Reviewed by Dean Jackson.

Source/WebCore:

We create a compositing layer for <embed> objects whose plug-in supports
the CoreAnimation rendering model. When the embed's renderer is not styled,
we can avoid allocating backing store for it.

Test: compositing/plugins/no-backing-store.html

* rendering/RenderLayerBacking.cpp:
(WebCore::isCompositedPlugin):
(WebCore::RenderLayerBacking::isSimpleContainerCompositingLayer):

LayoutTests:

Test for backing store on CoreAnimation-based plugins in various
configurations.

* compositing/plugins/no-backing-store-expected.txt: Added.
* compositing/plugins/no-backing-store.html: Added.

Identifier: 111401@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125101 268f45cc-cd09-0410-ab3c-d52691b4dbfc