[CSS Shaders] Invalid shaders should act as pass-through filters https://bugs.webkit.org/show_bug.cgi?id=93405
Reviewed by Dean Jackson.

Source/WebCore:

If the shader fails to apply then clearShaderResult will just copy the result of the previous filter
to the output of the current filter.

Test: css3/filters/custom/invalid-custom-filter-shader.html

* platform/graphics/filters/FECustomFilter.cpp:
(WebCore::FECustomFilter::platformApplySoftware):
(WebCore):
(WebCore::FECustomFilter::clearShaderResult):
(WebCore::FECustomFilter::applyShader):
* platform/graphics/filters/FECustomFilter.h:
(FECustomFilter):
* rendering/style/StyleCustomFilterProgram.h: The test was exposing the fact that if the shaders were referencing the same
file then StyleCustomFilterProgram will never complete the load. Having the same file for both the vertex and the fragment shader
cannot really work, because the shaders would not compile anyway, thus triggering an invalid shader.
I'm fixing it part of this change because the current test actually exposes that on Safari Mac builds. See the note in the LayoutTest/ChangeLog.
(WebCore::StyleCustomFilterProgram::notifyFinished):

LayoutTests:

Note that invalid-custom-filter-shader-expected.html is actually using another builtin filter,
just to make sure we trigger the whole FilterEffectRenderer on the images, otherwise the result will be
slightly different because of the conversion errors. Having another filter like grayscale(0) will trigger
the same precission errors on the result of the expected html.

* css3/filters/custom/invalid-custom-filter-shader-expected.html: Added.
* css3/filters/custom/invalid-custom-filter-shader.html: Added.
* css3/filters/resources/invalid-shader.vs: Added.


Identifier: 111423@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125128 268f45cc-cd09-0410-ab3c-d52691b4dbfc