Extract the definition of NativeImagePtr from ImageSource.h into NativeImagePtr.h. https://bugs.webkit.org/show_bug.cgi?id=93466
Patch by Huang Dongsung <luxtella@company100.net> on 2012-08-08
Reviewed by Adam Barth.

Image.h includes ImageSource.h in order to know the definition of NativeImagePtr.
But Image does not need to know the definition of ImageSource. So, this patch
extracts the definition of NativeImagePtr from ImageSource.h into NativeImagePtr.h.

This is a preparation for asynchronous image decoding. Async decoding
will need a callback function, which sends a NativeImagePtr to clients, and the
callback must know the definition of NativeImagePtr, not ImageSource.

No new tests, behavior is unchanged.

* GNUmakefile.list.am:
* Target.pri:
* WebCore.gypi:
* WebCore.vcproj/WebCore.vcproj:
* WebCore.xcodeproj/project.pbxproj:
* platform/graphics/BitmapImage.h:
* platform/graphics/Image.h:
* platform/graphics/ImageSource.h:
(WebCore):
* platform/graphics/NativeImagePtr.h: Added.
(WebCore):

Identifier: 111434@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125143 268f45cc-cd09-0410-ab3c-d52691b4dbfc