[CSS Regions] region-overflow: break still renders the content that does not fit in the last region. https://bugs.webkit.org/show_bug.cgi?id=92996
Reviewed by Julien Chaffraix.

Source/WebCore:

Flow content that follows the last break in the last region, if any, should not be rendered
for the regions with region-overflow: break.
http://www.w3.org/TR/css3-regions/#the-region-overflow-property

Modified the expected result for existing test fast/regions/region-overflow-break.html
to reflect the correct behavior.

* rendering/RenderRegion.cpp:
(WebCore::RenderRegion::regionOversetRect):

LayoutTests:

Modified expected result for original test to match the behavior described in the regions spec.

* fast/regions/region-overflow-break-expected.html:


Identifier: 111557@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125271 268f45cc-cd09-0410-ab3c-d52691b4dbfc